Generalizes Lightning acceleration.
This PR adds a flag `--accelerator` to the trainer and predictor. This
is then passed directly to the `Trainer` object. This object
automatically moves the model (etc.) to the chosen accelerator,
something we were doing explicitly beforehand. This required some
auditing of tensors created with `torch.zeros` and the like, but
extensive tests with Maori (no features) and Polish (has features), on
both training and prediction, all work on GPU without any problem now.
Once I figured out an appropriate auditing method (would you guess...it
was print-debugging...) I was able to solve #9 quickly.

Since the evaluator doesn't have any relevant state, I have not
bothered to move it either.

Should one fail to specify an accelerator, the default (CPU) is used but
users are warned about the presence of better accelerators (if
applicable).

While this has not come up yet, this also means the library now
naturally supports proprietary ASICs like TPU.

Closes #9: the transducer now works on GPU.

Closes #10: I have added information about accelerators to the README.