Adds automated LR tuning.
The PTL trainer supports two kinds of auto-tuning as an alternative to actual training:
one for batch size (which just tries to find the biggest value) and one for LR (which is a bit more sophisticated). Here, I have made the --auto_lr_find flag work.

If this flag is invoked, we run the tuner instead of the trainer's fit method, log the reuslt, and then quit. Otherwise, training proceeds as normal.