Pins numpy version to < 2.0.0.
NumPy 2 is the first major release in almost 20 years and has some breaking changes. These apparently have not been incorporated into various dependencies (PyTorch at least) and break our tests. I imagine this pin will be temporary and will add a TODO to check on it in a few months.