Adds automatic batch sizing.
If requested, this finds the max batch size, then computes the "optimal" batch size to use such that it, plus multiple gradient accumulation steps, is as large as possible
but does not exceed the max. The major features of the sizer are also exposed as flags.

The values are all logged and set, so that one can combine this with training, or one can simply do this without training using a trick like `--max_epochs 0` as documented in the README.

I tested this by taking a large transformer model (many encoder and decoder layers), requesting an enormous batch size of 2048. In this case (and following my GPU memory utilization bears this out) it succeeds at batches of size 1024 but fails at batches of 2048, so it sets batch size = 1024 and uses two gradient accumulation steps per update.

I also repeated this with `--max_epochs 0`, in which case it simply logs the same results from the finder, and then halts.