Fixes bug with features encoder.
The logic in `train.py` was to use a separate features encoder module if:

- features are present in the datamodule
- the model requires features to be separate from source (e.g., pointer-generators)
- and a separate features encoder module is requested.

This is incorrect; the logic should read:

- features are present in the datamodule, and
    - a separate features encoder module is requested, or
    - the model requires features to be separate from source

In the case where the model requires it but a specific features encoder module is not specified, we use the same module as the source encoder; e.g., a pointer-generator LSTM will have an LSTM source encoder and an LSTM features encoder.

The README was also out of date: the flags are `--source_encoder_arch` and `--features_encoder_arch` (was missing the `_arch` part).