Separates vocabularies for logging.
Current logic is such that source and target vocabularies (which are stored solely for logging and debugging purposes) are unioned together if --tied_embeddings. However, the purpose of these logging statements has nothing to do with embedding tying, so we modify them so that they are not combined except when building the actual vocabulary. We also remove the special symbols from the logging since they serve no purpose there.