Clears the cache after an OOM during sweeps.
As reported in #227, memory is not freed after an OOM during the sweep,
meaning that, usually, all subsequent sweeps also fail. We solve this
by:

1. Cleaning up the `train.train` API with a more friendly endpoint. This
has the desirable effect that there is no longer any need to keep
`train_wandb_sweep.py` "up to date" with trainer interface: it just
takes `args` and handles the rest.
2. wrap the call to `train.train` in a `try`/`except` with a `finally:` block that clears the CUDA cache.

Closes #227.