Adds docs to the transformer module.
I include the citation for pre-layer normalization.

I also move docs from __init__ to the class-level docs. Because of
symbol mangling it is really hard (reqiures metaprogramming) for the
user to programmatically access the docs for an __init__ method. Thus
the convention is to place them at the class level, a convention most
tooling understands. This should be done more widely but I'll limit
myself to this module for now.