Cleanup to schedulers.
We redefine schedulers to work at the epoch level. This seems to be
standard practice with a few exceptions.

* the inverse-square-root scheduler with warmup is now made to work at the epoch level. We also eliminate an unreachable condition, ensure the lambda-coefficient is never zero, rename flags etc.
* the linear decay scheduler, which we basically haven't used and didn't include in any of our hyperparameter work, is removed.

Note that the LightningCLI migration will make it possible for users to
use basically any scheduler in PyTorch, Lightning, or specified by the
user, so this is a temporary simplification to enable that migration.