Simplifies greedy decoding in the RNN
The length bounds for greedy decoding are as follows:

* If a target is provided, we need loss and maybe also are doing teacher
  forcing, and therefore do as many decoding steps as the longest
sequence in the target tensor.
* If no target is provided, then `max_target_length` acts as an upper
  bound. However, we can halt earlier if all sequences have decoded at
least one END.

This disjunction is currently not well encoded: we generate a vector
`finished` (tracking the presence of END in the sequences), and repeatedly update and test it, even in the case where a target is present. This PR fixes that: the vector is only generated, updated, and tested if no target is present.

We also update the method docs to reflect this and to draw the contrast
with beam decoding.