Doesn't use dropout inplace anymore.
This reverts an earlier change (#293) which used in-place dropout.

On my PC (Windows 11, WSL, older Nvidia GPU) in-place dropout throws the
inscrutable message:

```
RuntimeError: one of the variables needed for gradient computation has been modified by an inplace operation: [torch.cuda.FloatTensor [128, 1, 512]], which is output 0 of CudnnRnnBackward0, is at version 1; expected version 0 instead. Hint: enable anomaly detection to find the operation that failed to compute its gradient, with torch.autograd.set_detect_anomaly(True).
```

Tracing down this mysterious tensor of shape 128 x 1 x 512 I found that
it was the tensor input to dropout.

Note however that I avoid the expensive overwriting operation all the
same, so we get most of the same benefits as #293.