Avoids 0-to-a-negative exponent error.
When using the warmup inverse square root scheduler, if one specifies no
warmup, you will hit the following error:

```
    return self.decay_factor * step**-0.5
                               ~~~~^^~~~~
ZeroDivisionError: 0.0 cannot be raised to a negative power
```

This can simply be avoided by treating step as 1-based. We already do
this in the warmup clause, so we now do this in the warmed-up clause
too. I also slightly improve the comments since this is an important
point.