transducers: use learned decoder hidden state
This causes transducers to use the learned decoder hidden state rather than the encoder hidden state. This is parallel to #310 but for transducers rather than pointer-generators.

I believe because of the way this was initially set up, it is no-op in many cases and just involves deleting unreachable code. To confirm, I confirmed that performance on two languages and with GRUs and LSTMs was the same before and after.