Coerce pointer generator layer matching.
Currently the pointer-generator RNN implementation raises an error if
the number of encoder and decoder layers doesn't match. This is sensible
behavior in a library, but from the user perspective you might rather it
just fixes it. This also is a major issue for tuning the
pointer-generator; in the past both @Adamits and I have just made a
minor fork of the library to enforce the matching. What I propose here
is that we enforce it by setting the number of decoder layers to match
the number of encoder layers and then logging that to stderr.