Lowers version requirement on scipy.
Scipy 1.15 is the last that supports Python 3.10, which is still on
security release until next October. We downgrade the minimum
requirement in requirements.txt.

Scipy is only used for its implementation of Poisson sampling, which is
used to generate the "toy" test data. So it only needs to be listed in
requirements.txt; it's not a dependency for the library itself.