This PR removes an unused padding size feature.
The `PaddedTensor` object allows one to decide whether padding ought to
be determined automatically, or permits the user to specify a max
padding size. The latter feature is never used in the library---we
always use the automatic sizing---so I simply remove it, getting rid of
a if/else deep in the library.

I also do a driveby on some documentation.