Adds an informative error
This now fires when a features encoder is specified but
features are not.

E.g.: yoyodyne.models.hard_attention.Error: Features encoder enabled, but no feature column specified

I took the liberty of adding this separately to our five major classes
since the alternative, definining a `forward` method in the base class
and having all of the superclass call it like `super().forward(...)`
struck me as no more future-proof. (A new model designer could easily
forget to make this superclass call, or want/need a different signature
for `forward`.)

Reported by @Othergreengrasses, where the error was also inscrutable to
me until I put a bunch of debugging in. Seems like there's a need.