Adds OmegaConf parser support
Closes #232.

This enables OmegaConf as a parser of LightningCLI configuration files.

This has three major effects:

1. One can refer to other values in the same configuration file, e.g.,
   to set the nested hyperparameters of the encoders on the basis of
   the outer model/decoder values.
2. This eliminates the need for the LINK system, which is removed from
   the sweeping engine.
3. One can also define custom "resolutions" for mappings more complex
   than identity. Seeing only one such need, I add the ability to define
   one argument as a mathematical product of another.

This is extensively documented: example configuration files are
converted to use this convenient format, as are test configuration
files, and it is documented for use with the sweeping engine, where it
is terribly useful.

As a drive-by I improved default hidden layer and embedding sizes in the
tests, which are thus regenerated. I also rationalized some redundancies
in the base module; the dropout (probability and layer) and embedding size
are now stored in the base whereas other elements are derived as needed.