Adds new dataset objects
Closes #346.

The training dataset object, instead of being in memory, is now a
mmapped file. This adds a small amount of overhead at initialization,
because it has to compute the offsets, but this isn't so bad because 1)
we already had to do this and 2) it warms up the cache, so this should
behave as before if the file is small as far as the OS is concerned,
but it removes all practical limits on the size of a training file.
Most of this overhead is delegated to a separate data loader process,
which in turn delegates most of it to the kernel.

Datasets which don't require shuffling/random access are now implemented
iterably. This also removes all limits on file size and comes without
any meaningful penalties.